# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo App",
    "version": "1.0.0",
    "description": "API documentation for Todos App"
  },
  "servers": [
    {
      "url": "http://localhost:54321",
      "description": "Local server"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"],
                "example": {
                  "email": "user@example.com",
                  "password": "password123"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Error during registration"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"],
                "example": {
                  "email": "user@example.com",
                  "password": "password123"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully"
          },
          "400": {
            "description": "Error during login"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "Logout a user",
        "responses": {
          "200": {
            "description": "User logged out successfully"
          },
          "400": {
            "description": "Error during logout"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/todos": {
      "post": {
        "summary": "Create a new todo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" }
                },
                "required": ["title"],
                "example": {
                  "title": "Buy groceries"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo created successfully"
          },
          "400": {
            "description": "Error during todo creation"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get all todos for the authenticated user",
        "responses": {
          "200": {
            "description": "List of todos"
          },
          "400": {
            "description": "Error fetching todos"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/todos/{id}": {
      "put": {
        "summary": "Update a todo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "IsDone": { "type": "boolean" }
                },
                "required": ["title", "IsDone"],
                "example": {
                  "title": "Buy groceries",
                  "IsDone": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo updated successfully"
          },
          "400": {
            "description": "Error during todo update"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a todo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo deleted successfully"
          },
          "400": {
            "description": "Error during todo deletion"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/sessions": {
      "get": {
        "summary": "Get session logs for the authenticated user",
        "responses": {
          "200": {
            "description": "List of session logs"
          },
          "400": {
            "description": "Error fetching session logs"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}